+++ src/Base/Foundation/Plugin.php
@@ -25,7 +25,7 @@ class Plugin
      *
      * @var string VERSION
      */
-    const VERSION = '2.0.17';
+    const VERSION = '2.0.18';
 
     /**
      * Path to the root of the plugin.